Support IVRSystem_009, IVRCompositor_009, IVROverlay_007
Needed for Project Cars 2.
It seems that there were some breaking API changes between OpenVR version 0.9.12
and 1.0, so the openvr bindings generator is adjusted to handle this. Types that
are incompatible with 1.0 are moved into their own module (vr_0_9_12) and
implementing the interfaces requires using the appropriate types.

Part of #129